Move `DayRangeValidationProps` type back to internal models
