+++ packages/x-date-pickers-pro/src/internals/models/index.ts
@@ -1,3 +1,4 @@
+export * from './dateRange';
 export * from './dateTimeRange';
 export * from './timeRange';
 export * from './rangePickerProps';
