+++ packages/x-date-pickers-pro/src/models/index.ts
@@ -1,4 +1,4 @@
-export type { UseDateRangeFieldProps } from './dateRange';
+export * from './dateRange';
 export * from './fields';
 export * from './range';
 export * from './validation';
